Increased difficulty of median problem, added lots more problems of all difficulties

+++ src/Mild2.java
@@ -6,7 +6,7 @@
 public class Mild2 {
 
     public int[] a;
-    public int mean;
+    public double mean; // double instead of int because mean might be a decimal
 
     public static void main(String[] args) {
         Mild2 practice = new Mild2();
@@ -15,6 +15,7 @@ public class Mild2 {
     public Mild2() {
 
         a = new int[20];
+
         for (int i = 0; i < a.length; i++) {
             a[i] = (int)(Math.random()*100);
         }
